+++ python/mirage/config.py
@@ -24,8 +24,9 @@ from mirage.accessor.s3 import S3Config
 from mirage.cache.file.config import CacheConfig, RedisCacheConfig
 from mirage.cache.index.config import IndexConfig, RedisIndexConfig
 from mirage.resource.registry import build_resource
-from mirage.runtime.base import Runtime, ScriptSource
+from mirage.runtime.base import Runtime
 from mirage.runtime.table import build_runtime
+from mirage.runtime.types import ScriptSource
 from mirage.types import (KERNEL_BACKENDS, CommandSafeguard, ConsistencyPolicy,
                           MountBackend, MountMode)
 from mirage.workspace.mount.spec import Mount
