+++ python/tests/runtime/js/test_quickjs.py
@@ -20,9 +20,9 @@ import pytest
 
 from mirage import MountMode, Workspace
 from mirage.resource.ram import RAMResource
-from mirage.runtime.base import RunArgs
 from mirage.runtime.js import QuickJsRuntime
 from mirage.runtime.js.quickjs import QUICKJS_HOME_ENV
+from mirage.runtime.types import RunArgs
 
 
 def _home_dir() -> str | None:
