+++ typescript/packages/core/src/workspace/executor/python/runtimes/pyodide.ts
@@ -12,17 +12,17 @@
 // limitations under the License.
 // ========= Copyright 2026 @ Strukto.AI All Rights Reserved. =========
 
-import {
-  EvalError,
-  Runtime,
-  type EvalResult,
-  type EvalStatus,
-  type Evaluator,
-  type EvalValue,
-  type RunArgs,
-  type RunResult,
-  type RuntimeOptions,
-} from '../../runtime.ts'
+import { Runtime } from '../../runtime.ts'
+import { EvalError } from '../../runtime_errors.ts'
+import type { Evaluator } from '../../runtime_mixin.ts'
+import type {
+  EvalResult,
+  EvalStatus,
+  EvalValue,
+  RunArgs,
+  RunResult,
+  RuntimeOptions,
+} from '../../runtime_types.ts'
 import { loadPyodideRuntime, type PyodideInterface } from '../loader.ts'
 import {
   createMirageBridge,
