+++ typescript/packages/core/src/workspace/executor/runtime.ts
@@ -14,142 +14,7 @@
 
 import type { BridgeDispatchFn } from './python/mirage_bridge.ts'
 import { ScriptSource, type RouteScript } from './route/types.ts'
-
-/** One interpreter execution request, language-agnostic. */
-export interface RunArgs {
-  code: string
-  args: string[]
-  env: Record<string, string>
-  stdin: Uint8Array | null
-  /**
-   * Interpreter-level switches parsed by the command's spec (e.g. js
-   * module mode). Each runtime reads its own switches and ignores the
-   * rest.
-   */
-  flags?: Record<string, unknown>
-}
-
-/** Outcome of one interpreter execution. */
-export interface RunResult {
-  stdout: Uint8Array
-  /** Captured standard error, null when empty (mirrors Python). */
-  stderr: Uint8Array | null
-  exitCode: number
-}
-
-/**
- * The value contract of eval: never richer than JSON plus bytes, so
- * any evaluator (in-process or remote over a serialized transport)
- * can carry it, in either direction (inputs in, verdict out).
- */
-export type EvalValue =
-  | null
-  | boolean
-  | number
-  | string
-  | Uint8Array
-  | EvalValue[]
-  | { [key: string]: EvalValue }
-
-/**
- * "incomplete" is console semantics: the source needs a continuation
- * line (session mode only). "exit" is an explicit exit() call.
- */
-export type EvalStatus = 'complete' | 'incomplete' | 'exit'
-
-/**
- * Outcome of one evaluation (mirrors Python's EvalResult).
- *
- * One-shot mode rejects with EvalError on any failure, so a returned
- * result is always a success. Session (console) mode is a transcript:
- * a failing snippet comes back as a result too (its traceback on
- * stderr, a nonzero exitCode), because a console reports errors and
- * keeps going.
- */
-export interface EvalResult {
-  /**
-   * The program's last expression. In-process evaluators return it
-   * directly; remote ones return what the transport could carry.
-   * Session (console) mode may report null when the evaluator only
-   * streams output.
-   */
-  value: EvalValue
-  /** Output the program printed while running. */
-  stdout: Uint8Array
-  /** Captured standard error, null when empty. */
-  stderr: Uint8Array | null
-  /** 0 outside session mode; a console snippet's exit. */
-  exitCode: number
-  /** Console verdict; always "complete" outside session mode. */
-  status: EvalStatus
-}
-
-/**
- * An evaluation that could not produce a value. The message carries
- * the evaluator's own diagnostics (a traceback, a transport failure,
- * a non-serializable result); `syntax` is true when the program
- * failed to parse, so callers can distinguish "bad script" from
- * "script raised".
- */
-export class EvalError extends Error {
-  readonly syntax: boolean
-
-  constructor(message: string, options: { syntax?: boolean; cause?: unknown } = {}) {
-    super(message, options.cause !== undefined ? { cause: options.cause } : {})
-    this.name = 'EvalError'
-    this.syntax = options.syntax ?? false
-  }
-}
-
-/**
- * The evaluator capability: named inputs in, a value out (Python's
- * EvaluatorMixin). A Runtime that also implements this can evaluate
- * expressions, which is what the routing policy engine and the repl
- * consume; process-only runtimes never implement it and are never
- * asked to evaluate. The contract promises the shape, not value
- * fidelity: inputs and the returned value stay within EvalValue so
- * any transport can carry them, and errors surface as the
- * evaluator's own diagnostics wrapped in EvalError.
- */
-export interface Evaluator {
-  /**
-   * Evaluate one program and return its last expression. `inputs`
-   * bind as globals in the evaluator's own idiom; a `session` id
-   * keeps state alive per id (console semantics), absent evaluates
-   * one-shot.
-   */
-  eval(
-    code: string,
-    opts?: { inputs?: Record<string, EvalValue>; session?: string },
-  ): Promise<EvalResult>
-}
-
-/** Whether this runtime carries the evaluator capability. */
-export function isEvaluator(runtime: Runtime): runtime is Runtime & Evaluator {
-  return typeof (runtime as Partial<Evaluator>).eval === 'function'
-}
-
-/** Constructor options every runtime accepts (a yaml entry's keys). */
-export interface RuntimeOptions<C extends object = Record<string, unknown>> {
-  /**
-   * Commands this runtime claims, overriding the class default; ["*"]
-   * claims every line for a runsLines runtime.
-   */
-  captures?: readonly string[]
-  /**
-   * The runtime's implementation knobs (a yaml entry's `config`
-   * block), coerced against the runtime's own key list so a field the
-   * runtime does not have fails loud.
-   */
-  config?: C
-  /**
-   * Per-line admission script for the routing ladder, answering "do I
-   * want this line": a function taking a RouteContext, or a
-   * config-borne ScriptSource. Absent = always willing. Policy, not
-   * capability: it can only refuse lines the captures already allow.
-   */
-  script?: RouteScript
-}
+import type { RunArgs, RunResult, RuntimeOptions } from './runtime_types.ts'
 
 /**
  * A constructor's config option as the runtime's own config, mirroring
